Moved enums to respective configs, fix display logic when both are visible
